+++ test/sass-frameworks-spec.js
@@ -8,6 +8,7 @@ var fs = require('fs');
 
 describe('Sass Frameworks', function () {
 	var srcPath = "resources/";
+	var helpersPath = "helpers/";
 
 	describe('when Bootstrap-Sass is selected', function () {
 		var answers = require('../test_helpers/prompt-answer-factory')({
@@ -101,6 +102,101 @@ describe('Sass Frameworks', function () {
 
 		it('adds comment to _vars.scss', function () {
 			assert.fileContent(srcPath + 'scss/global/_vars.scss', /BOURBON/);
+		});
+	});
+
+	describe('when Lost Grid is selected', function () {
+		var answers = require('../test_helpers/prompt-answer-factory')({
+			"cssLibs": [
+				"lost-grid"
+			]
+		});
+
+		beforeEach(function (done) {
+			helpers.run(path.join(__dirname, '../generators/app'))
+				.inDir(path.join(__dirname, 'tmp'))
+				.withOptions({
+					'skip-install': true,
+					'skip-welcome-message': true
+				})
+				.withPrompts(answers)
+				.on('end', done);
+		});
+
+		it('adds the package to package.json', function () {
+			assert.fileContent('package.json', /lost/);
+		});
+
+		it('adds the package to the task file', function () {
+			assert.fileContent('package.json', /lost/);
+		});
+
+		it('adds comment to _vars.scss', function () {
+			assert.fileContent(srcPath + 'scss/global/_vars.scss', /LOST/);
+		});
+	});
+
+	describe('when Lost Grid and Grunt is selected', function () {
+		var answers = require('../test_helpers/prompt-answer-factory')({
+			"cssLibs": [
+				"lost-grid"
+			]
+		});
+
+		beforeEach(function (done) {
+			helpers.run(path.join(__dirname, '../generators/app'))
+				.inDir(path.join(__dirname, 'tmp'))
+				.withOptions({
+					'skip-install': true,
+					'skip-welcome-message': true
+				})
+				.withPrompts(answers)
+				.on('end', done);
+		});
+
+		it('adds the package to package.json', function () {
+			assert.fileContent('package.json', /grunt-postcss/);
+		});
+
+		it('adds the package to the task and watch file', function () {
+			assert.fileContent('package.json', /grunt-postcss/);
+			assert.fileContent(helpersPath + '_grunt/postcss.js', /lost/);
+			assert.fileContent(helpersPath + '_grunt/watch.js', /postcss:dev/);
+		});
+
+		it('adds the task to the Gruntfile file', function () {
+			assert.fileContent('Gruntfile.js', /postcss:dev/);
+		});
+	});
+
+	describe('when Lost Grid and Gulp is selected', function () {
+		var answers = require('../test_helpers/prompt-answer-factory')({
+			'taskRunner': [
+				'gulp'
+			],
+			"cssLibs": [
+				"lost-grid"
+			]
+		});
+
+		beforeEach(function (done) {
+			helpers.run(path.join(__dirname, '../generators/app'))
+				.inDir(path.join(__dirname, 'tmp'))
+				.withOptions({
+					'skip-install': true,
+					'skip-welcome-message': true
+				})
+				.withPrompts(answers)
+				.on('end', done);
+		});
+
+		it('adds the package to package.json', function () {
+			assert.fileContent('package.json', /gulp-postcss/);
+		});
+
+		it('adds the package to the task file', function () {
+			assert.fileContent('package.json', /gulp-postcss/);
+			assert.fileContent(helpersPath + '_gulp/styles.js', /lost/);
 		});
 	});