+++ test/veams-packages-spec.js
@@ -10,7 +10,7 @@ describe('veams extensions', function () {
 	var helperPath = "helpers/";
 	var srcPath = "resources/";
 
-	describe('when pg methodology is installed', function () {
+	describe('when veams methodology is installed', function () {
 		var answers = require('../test_helpers/prompt-answer-factory')({
 			"templateEngine": "assemble",
 			"veamsPackages": [
@@ -32,16 +32,16 @@ describe('veams extensions', function () {
 		it('creates READMEs', function () {
 			var expected = [
 				srcPath + "templating/layouts/README.md",
-				srcPath + "templating/partials/README.md",
 				srcPath + "templating/partials/blocks/README.md",
-				srcPath + "templating/partials/components/README.md"
+				srcPath + "templating/partials/components/README.md",
+				srcPath + "templating/partials/utilities/README.md"
 			];
 			assert.file(expected);
 		});
 
 	});
 
-	describe('when veams-scss is installed', function () {
+	describe('when veams-sass is installed', function () {
 		var answers = require('../test_helpers/prompt-answer-factory')({
 			"templateEngine": "assemble",
 			"veamsPackages": [
@@ -121,30 +121,6 @@ describe('veams extensions', function () {
 		});
 	});
 
-	describe('when veams-components is installed', function () {
-		var answers = require('../test_helpers/prompt-answer-factory')({
-			"templateEngine": "assemble",
-			"veamsPackages": [
-				"veamsComponents"
-			]
-		});
-
-		beforeEach(function (done) {
-			helpers.run(path.join(__dirname, '../generators/app'))
-					.inDir(path.join(__dirname, 'tmp'))
-					.withOptions({
-						'skip-install': true,
-						'skip-welcome-message': true
-					})
-					.withPrompts(answers)
-					.on('end', done);
-		});
-
-		it('bower.json contains reference', function () {
-			assert.fileContent('bower.json', /veams-components/);
-		});
-	});
-
 	describe('when veams is not installed', function () {
 		var answers = require('../test_helpers/prompt-answer-factory')({
 			"templateEngine": "assemble",
@@ -163,7 +139,7 @@ describe('veams extensions', function () {
 		});
 
 		it('bower.json contains no references', function () {
-			assert.noFileContent('bower.json', /veams-components|veams-sass|veams-js/);
+			assert.noFileContent('bower.json', /veams-sass|veams-js/);
 		});
 	});